backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-CORS
Flask-Migrate
PyMySQL
requests
python-dotenv
